% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\name{air_base}
\alias{air_base}
\title{Get airtable base object}
\usage{
air_base(base, tables)
}
\arguments{
\item{base}{Airtable base}

\item{tables}{Table names in the airtable base (character vector)}
}
\value{
Airtable base object
}
\description{
Creates airtable object with tables and functions
}

