% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_create_description_table}
\alias{air_create_description_table}
\title{Create the descriptive metadata table for the base}
\usage{
air_create_description_table(
  base,
  description,
  table_name = "Description",
  field_descriptions = NA,
  type = "singleLineText",
  options = NA
)
}
\arguments{
\item{base}{String. Base id}

\item{description}{Data frame. Description from air_get_base_description* or air_generate_base_description}

\item{table_name}{String. Name of description table}

\item{field_descriptions}{Character vector. Descriptions of metadata table fields. If NA, DCMI terms will be used where possible.}

\item{type}{Character vector. Column types for metadata table fields. see \url{https://airtable.com/developers/web/api/field-model}}

\item{options}{Data frame. Options for fields in metadata table.}
}
\value{
List. Outputs from creating the table and inserting the records
}
\description{
Create the descriptive metadata table for the base
}
\details{
DCMI terms can be found here \url{https://www.dublincore.org/specifications/dublin-core/dcmi-terms/}
}
\examples{
\dontrun{
base = "appVjIfAo8AJlfTkx"
table_name= "description"

description <- air_generate_base_description(title = "Example Base",
                                             creator = "Collin Schwantes")

air_create_description_table(base,description,table_name)
}

}
