% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_create_field}
\alias{air_create_field}
\title{Create a new field in a table}
\usage{
air_create_field(
  base,
  table_id,
  name,
  description = NA,
  type = "singleLineText",
  options = NA
)
}
\arguments{
\item{base}{String. Base id}

\item{table_id}{String. Table id. Can be found using \code{air_get_schema}}

\item{name}{String. Name of the field}

\item{description}{String. Description of the field}

\item{type}{String. Type of field. See https://airtable.com/developers/web/api/field-model}

\item{options}{Data frame. See https://airtable.com/developers/web/api/field-model}
}
\value{
description of newly created field as a list
}
\description{
See https://airtable.com/developers/web/api/create-field
}
\examples{
\donotrun{
base_schema <- air_get_schema(base)

base_schema$tables

air_create_field(base,table_id = base_schema$tables$id[[4]],
                 name = "Has Nucleics",
                 description = "Logical. Does this planet have nucleics?",
                 type = "checkbox",
                 options = list(
                   list(
                     "color"= "greenBright",
                     "icon"= "check"
                   )
                 )
)
}

}
