% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_create_metadata_table}
\alias{air_create_metadata_table}
\title{Create a new structural metadata table in the base}
\usage{
air_create_metadata_table(
  base,
  meta_data,
  table_name = "Meta Data",
  field_descriptions = NA,
  type = "singleLineText",
  options = NA
)
}
\arguments{
\item{base}{String. Base id}

\item{meta_data}{Data frame. Contains metadata records. From air_generate_metadata*}

\item{table_name}{String. name of the metadata table. default is "Meta Data"}

\item{field_descriptions}{Character vector. Descriptions of metadata table fields}

\item{type}{Character vector. Column types for metadata table fields. see https://airtable.com/developers/web/api/field-model}

\item{options}{Data frame. Options for fields in metadata table.}
}
\value{
List with outcome from creating the table and inserting the records
}
\description{
Create a new structural metadata table in the base
}
\examples{
\dontrun{
# set base id
base <- "appXXXXXXXX"
# create metadata from api
metadata  <- air_generate_metadata_from_api(base)
# add Meta Data table to base -- will not work if base already has a metadata
# table
log <- air_create_metadata_table(base,metadata)

}


}
