% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_create_table}
\alias{air_create_table}
\title{A function to create new tables in a base}
\usage{
air_create_table(base, table_list)
}
\arguments{
\item{base}{String. ID for the base}

\item{table_list}{List. see \code{air_table_template}}
}
\value{
Data frame of table schema
}
\description{
Takes a list object with appropriate arguments (see \code{air_table_template})
converts it to JSON then adds it to the specified base.
}
\note{
See https://airtable.com/developers/web/api/create-table
}
