% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\name{air_delete}
\alias{air_delete}
\title{Delete a record}
\usage{
air_delete(base, table_name, record_id)
}
\arguments{
\item{base}{Airtable base}

\item{table_name}{Table name}

\item{record_id}{Id of the record to be deleted}
}
\description{
Deletes a record and returns the deleted record id if the call
succeeded.
}
