% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_download_attachments.R
\name{air_download_attachments}
\alias{air_download_attachments}
\title{Download Airtable file attachments}
\usage{
air_download_attachments(x, field, dir_name = "downloads", ...)
}
\arguments{
\item{x}{Data frame. Output from air_get or fetch_all.}

\item{field}{String. Name of field with file attachments in base}

\item{dir_name}{String. Where should files be downloaded to?
Will create the folder if it does not exist. Folders created are recursively.}

\item{...}{reserved for additional arguments.}
}
\value{
Returns x with an additional field called attachment_file_paths
}
\description{
Download an attachment stored in air tables. Returns original dataframe
with an additional field called attachment_file_paths. The attachment_file_paths
field is of class list so it can handle multiple attachments per record. File
paths are prepended with record ids so that all file names are unique.
}
\examples{
\dontrun{

base <- "appXXXXXXXXX"
table_name <- "Table With Attachments"

table_original  <- air_get(base,table_name)

table_with_file_paths <- air_download_attachments(x = table_with_attachments,
                        field = "attachment_field",
                        dir_name = "downloads")

table_with_file_paths$attachment_file_paths

}

}
