% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_dump}
\alias{air_dump}
\title{Dump all tables from a base into R}
\usage{
air_dump(base, metadata, description = NULL, add_missing_fields = TRUE)
}
\arguments{
\item{base}{String. ID for your base from Airtable. Generally 'appXXXXXXXXXXXXXX'}

\item{metadata}{Data.frame.Data frame with structural metadata - describes relationship between tables and fields.}

\item{description}{Data.frame. Data frame with descriptive metadata - describes whats in your base and who made it.
Can be left as NULL if base already contains a table called description.}

\item{add_missing_fields}{Logical. Should fields described in the metadata data.frame be added to corresponding tables?}
}
\value{
List of data.frames. All tables from metadata plus the
description and metadata tables.
}
\description{
Dump all tables from a base into R
}
