% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_dump_to_json}
\alias{air_dump_to_json}
\title{Dump all tables from a base into json files}
\usage{
air_dump_to_json(
  base,
  metadata,
  description = NULL,
  output_dir = "outputs",
  overwrite = FALSE
)
}
\arguments{
\item{base}{String. ID for your base from Airtable. Generally 'appXXXXXXXXXXXXXX'}

\item{metadata}{Data.frame.Data frame with structural metadata - describes relationship between tables and fields.}

\item{description}{Data.frame. Data frame with descriptive metadata - describes whats in your base and who made it.
Can be left as NULL if base already contains a table called description}

\item{output_dir}{String. Where should json files be saved?}

\item{overwrite}{Logical. If data are not unique, should files be overwritten?}
}
\value{
List of data.frames. All tables from metadata plus the
description and metadata tables.
}
\description{
Essentially air_get without converting to Rs. Does not add fields with empty
values.
}
