% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_fields_df_template}
\alias{air_fields_df_template}
\title{Template for for creating a table from a dataframe}
\usage{
air_fields_df_template(name, description, type, options = NA)
}
\arguments{
\item{options}{}
}
\description{
Template for for creating a table from a dataframe
}
