% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_fields_list_from_template}
\alias{air_fields_list_from_template}
\title{Convert field data frame to list}
\usage{
air_fields_list_from_template(df)
}
\arguments{
\item{df}{Data frame. From air_fields_df_template}
}
\value{
List. Structured for easy parsing into JSON
}
\description{
Converts the field data frame to a list of easier translation to JSON
}
\examples{
\dontrun{
base <- "appQ94sELAtFnXPxx"

base_schema <- air_get_schema(base)

tables<- base_schema$tables

field_names <- c("Planet","Chapter","Book", "Known Inhabitants")

field_desc <- c("Name of planet in Foundation Series",
                "Chapters where planet is referenced",
                "Books where planet is referenced",
                "Characters mentioned as living on or being from that planet")

field_types <- c("singleLineText",rep("multipleRecordLinks",3))

field_options <- c(NA,list(
  list(
    linkedTableId = tables[tables$name == "Chapter","id"]
  )
),
list(
  list(
    linkedTableId = tables[tables$name == "Book","id"]
  )
),
list(
  list(
    linkedTableId = tables[tables$name == "Character","id"]
  )
)
)

field_df <- air_fields_df_template(name = field_names,
                                  description = field_desc,
                                  type = field_types,
                                  options = field_options)

fields_list <- air_fields_list_from_template(df = fields_df)

}
}
