% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_generate_base_description}
\alias{air_generate_base_description}
\title{Generate descriptive metadata}
\usage{
air_generate_base_description(
  title = NA,
  primary_contact = NA,
  email = NA,
  base_description = NA,
  ...
)
}
\arguments{
\item{title}{String. Title is a property that refers to the name or names by
which a resource is formally known.}

\item{primary_contact}{String.  Person or entity primarily responsible for
making the content of a resource}

\item{email}{String. Email of primary_contact}

\item{base_description}{String. This property refers to the description of
the content of a resource. The description is a potentially rich source of
indexable terms and assist the users in their selection of an appropriate
resource.}

\item{...}{String. Additional descriptive metadata elements. See details.
Additional elements can be added as name pair values e.g.
\code{license = "CC BY 4.0", is_part_of = "https://doi.org/10.48321/MyDMP01"}}
}
\value{
data.frame with descriptive metadata
}
\description{
Creates a data.frame that describes the base.
}
\details{
See  \href{https://www.dublincore.org/resources/userguide/creating_metadata/}{dublin core} for inspiration about additional attributes.
}
\examples{

air_generate_base_description(title = "My Awesome Base" ,
 primary_contact= "Base Creator/Maintainer",
 email = "email@example.com",
 base_description = "This base is used to contain my awesome data
 from a project studying XXX in YYY. Data in the base were collected
 from 1900-01-01 to 1990-01-01 by researchers at Some Long Term Project.",
 is_part_of = "https://doi.org/10.48321/MyDMP01",
 is_part_of = "https://doi.org/10.5072/zenodo_sandbox.1062705"
 )

}
