% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_generate_metadata}
\alias{air_generate_metadata}
\title{Generated Metadata from table names}
\usage{
air_generate_metadata(base, table_names, limit = 1)
}
\arguments{
\item{base}{String. ID for your base from Airtable. Generally 'appXXXXXXXXXXXXXX'}

\item{table_names}{Vector of strings. The names of your tables. eg c("table 1","table 2", etc.)}

\item{limit}{Number from 1-100. How many rows should we pull from each table to create the metdata?
Keep in mind that the airtable api will not return fields with "empty" values - "", false, or [].
Code runs faster if fewer rows are pulled.}
}
\value{
data.frame with structural metadata.
}
\description{
Generates a structural metadata table - the metadata that
describes how tables and fields fit together. Does not
include field types.
}
\details{
For information about creating metadata tables in your base see the
\href{https://ecohealthalliance.github.io/eha-ma-handbook/8-airtable.html#managing-data}{EHA MA Handbook}
}
