% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_generate_metadata_from_api}
\alias{air_generate_metadata_from_api}
\title{Generate structural metadata from the api}
\usage{
air_generate_metadata_from_api(
  base,
  metadata_table_name = "Meta Data",
  include_metadata_table = FALSE
)
}
\arguments{
\item{base}{String. Base id}

\item{metadata_table_name}{String. Name of exisiting structural metadata table if it exists}

\item{include_metadata_table}{Logical. Should the structural metadata table be included in the metadata?}
}
\value{
A data frame with metadata
}
\description{
Generate structural metadata from the api
}
\examples{

\dontrun{

base <- "appXXXXXXXX"
metadata  <- air_generate_metadata_from_api(base)

}
}
