% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\name{air_get}
\alias{air_get}
\title{Get a list of records or retreive a single}
\usage{
air_get(base, table_name, record_id = NULL, limit = NULL, offset = NULL,
  view = NULL, sortField = NULL, sortDirection = NULL,
  combined_result = TRUE)
}
\arguments{
\item{base}{Airtable base}

\item{table_name}{Table name}

\item{record_id}{(optional) Use record ID argument to retrieve an existing
record details}

\item{limit}{(optional) A limit on the number of records to be returned.
Limit can range between 1 and 100.}

\item{offset}{(optional) Page offset returned by the previous list-records
call. Note that this is represented by a record ID, not a numerical offset.}

\item{view}{(optional) The name or ID of the view}

\item{sortField}{(optional) The field name to use for sorting}

\item{sortDirection}{(optional) "asc" or "desc". The sort order in which the
records will be returned. Defaults to asc.}

\item{combined_result}{If TRUE (default) all data is returned in the same data
frame. If FALSE table fields are returned in separate \code{fields} element.}
}
\value{
A data frame with records or a list with record details if
  \code{record_id} is specified.
}
\description{
You can retrieve records in an order of a view by providing the name or ID of
the view in the view query parameter. The results will include only records
visible in the order they are displayed.
}
