% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\name{air_get_all}
\alias{air_get_all}
\title{Get all records from table}
\usage{
air_get_all(
  base,
  table_name,
  view = NULL,
  sortField = NULL,
  sortDirection = NULL
)
}
\arguments{
\item{base}{Airtable base}

\item{table_name}{Table name}

\item{view}{(optional) The name or ID of the view}

\item{sortField}{(optional) The field name to use for sorting}

\item{sortDirection}{(optional) "asc" or "desc". The sort order in which the
records will be returned. Defaults to asc.}
}
\value{
A data frame with records
}
\description{
You can retrieve all records from a table. The results will include only records
visible in the order they are displayed.
}
