% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_get_attachments.R
\name{air_get_attachments}
\alias{air_get_attachments}
\title{Get Airtable file attachments}
\usage{
air_get_attachments(
  base,
  table_name,
  field,
  download_file = FALSE,
  dir_name = "downloads",
  extract_type = "excel",
  extract_field = "excel_extract",
  ...
)
}
\arguments{
\item{base}{String. ID for the base or app to be fetched}

\item{table_name}{String. Name of the table to be fetched from the base}

\item{field}{String. Name of field with file attachments in base}

\item{download_file}{Logical. Should files be downloaded?}

\item{dir_name}{String. Where should files be downloaded to?
Will create the folder if it does not exist.}

\item{extract_type}{String. File type to be extracted.
Should be one of: excel}

\item{extract_field}{String. Name of extract field that will be created}

\item{...}{Additional arguments to pass to \code{air_get}}
}
\value{
named list of data frames
}
\description{
Get an attachment stored in air tables. For excel files, returns a named list.
}
