% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_get_base_description_from_table}
\alias{air_get_base_description_from_table}
\title{Get base description from table}
\usage{
air_get_base_description_from_table(base, table_name)
}
\arguments{
\item{base}{String. ID for your base from Airtable. Generally 'appXXXXXXXXXXXXXX'}

\item{table_name}{String. Name of descriptive metadata table - the metadata that
describes the base and provides attribution}
}
\value{
data.frame with descriptive metadata.
}
\description{
Pull a table that has descriptive metadata.
Requires the following fields:
"title","primary_contact","email","base_description"
}
