% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_get_enterprise.R
\name{air_get_enterprise}
\alias{air_get_enterprise}
\title{Get data from the enterprise endpoint}
\usage{
air_get_enterprise(path, id, params = NULL)
}
\arguments{
\item{path}{String. What part of the metadata endpoint are you querying? 
See \url{https://airtable.com/developers/web/api/introduction}}

\item{id}{String. ID of item to be queried. e.g. workspace id or enterprise id.
If id = NULL then the id will be omitted from the request url.}

\item{params}{Optional. List. List of parameters}
}
\value{
list from jsonlite
}
\description{
Get data from the enterprise endpoint
}
