% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_from_url.R
\name{air_get_id_from_url}
\alias{air_get_id_from_url}
\alias{air_get_base_id_from_url}
\alias{air_get_table_id_from_url}
\alias{air_get_view_id_from_url}
\alias{air_get_record_id_from_url}
\title{Get an ID from a URL}
\usage{
air_get_id_from_url(url, pattern, id_type, split_pattern = "/|\\\\?")

air_get_base_id_from_url(url, pattern = "^app\\\\w{13}")

air_get_table_id_from_url(url, pattern = "^tbl\\\\w{13}")

air_get_view_id_from_url(url, pattern = "^viw\\\\w{13}")

air_get_record_id_from_url(url, pattern = "^rec\\\\w{13}")
}
\arguments{
\item{url}{String. A url generated by airtable}

\item{pattern}{String. A regex pattern for identifying the type of id you would like to get}

\item{id_type}{String. One of base_id, table_id, view_id, or record_id.}

\item{split_pattern}{String. Where should the URL be split? default is forward slashes "/" and questionmarks "?"}
}
\value{
String
}
\description{
General function for parsing airtable URLs to find base, table, view, or record id's
}
\section{Functions}{
\itemize{
\item \code{air_get_base_id_from_url()}: Get the base id

\item \code{air_get_table_id_from_url()}: Get the table id

\item \code{air_get_view_id_from_url()}: Get the view id

\item \code{air_get_record_id_from_url()}: Get the record id

}}
\examples{

url <- "https://airtable.com/apphDEokVZ9gvPNFk/tblaKC1ADBafoHVXN/viwteUgD7vaMBruHR/recMzdoM43RVRWybD?blocks=hide"

# General function for parsing url components
air_get_id_from_url(url, '^app',id_type = "base_id")
# Get different components
air_get_base_id_from_url(url)
air_get_table_id_from_url(url)
air_get_view_id_from_url(url)
air_get_record_id_from_url(url)

}
