% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_get_json.R
\name{air_get_json}
\alias{air_get_json}
\title{Get a list of records or retrieve a single record as JSON}
\usage{
air_get_json(
  base,
  table_name,
  record_id = NULL,
  limit = NULL,
  offset = NULL,
  view = NULL,
  fields = NULL,
  sortField = NULL,
  sortDirection = NULL,
  combined_result = TRUE,
  pretty = FALSE
)
}
\arguments{
\item{base}{Airtable base}

\item{table_name}{Table name}

\item{record_id}{(optional) Use record ID argument to retrieve an existing
record details}

\item{limit}{(optional) A limit on the number of records to be returned.
Limit can range between 1 and 100.}

\item{offset}{(optional) Page offset returned by the previous list-records
call. Note that this is represented by a record ID, not a numerical offset.}

\item{view}{(optional) The name or ID of the view}

\item{fields}{(optional) Only data for fields whose names are in this list
will be included in the records. If you don't need every field, you can use}

\item{sortField}{(optional) The field name to use for sorting}

\item{sortDirection}{(optional) "asc" or "desc". The sort order in which the
records will be returned. Defaults to asc.}

\item{combined_result}{If TRUE (default) all data is returned in the same data.
If FALSE table fields are returned in separate \code{fields} element.}

\item{pretty}{Logical. Should JSON be returned in human readable form?
this parameter to reduce the amount of data transferred.}
}
\value{
A data frame with records or a list with record details if
  \code{record_id} is specified.
}
\description{
Returns JSON objects from GET requests
}
