% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_get_metadata_from_table}
\alias{air_get_metadata_from_table}
\title{Pull the metadata table from airtable}
\usage{
air_get_metadata_from_table(base, table_name)
}
\arguments{
\item{base}{String. ID for your base from Airtable. Generally 'appXXXXXXXXXXXXXX'}

\item{table_name}{String. Name of structural metadata table - the metadata that
describes how tables and fields fit together.}
}
\value{
data.frame with metadata table
}
\description{
For information about creating metadata tables in your base see the
\href{https://ecohealthalliance.github.io/eha-ma-handbook/8-airtable.html#managing-data}{EHA MA Handbook}
}
\details{
Requires the following fields: table_name, field_name
}
