% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_list_bases}
\alias{air_list_bases}
\title{Get list of bases for an Token}
\usage{
air_list_bases(request_url = "https://api.airtable.com/v0/meta/bases")
}
\value{
list
}
\description{
Get list of bases for an Token
}
