% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\name{air_make_json}
\alias{air_make_json}
\title{Make JSON for API}
\usage{
air_make_json(
  base,
  table_name,
  record_data,
  record_id = NULL,
  method = "POST",
  typecast = TRUE
)
}
\arguments{
\item{base}{String. Base in airtable}

\item{table_name}{String. Table in airtable}

\item{record_data}{Dataframe, list, or vector. Data to be converted to JSON}

\item{record_id}{String or vector of strings. Records to be manipulated}

\item{method}{String. "PATCH" is necessary for \code{air_update}}

\item{typecast}{Logical. Should the typecast option be TRUE or FALSE? Typecast
allows you to add new options to select type fields.}
}
\value{
JSON with record data
}
\description{
Make JSON that is compatible with the Airtable API.
}
