% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\name{air_make_request}
\alias{air_make_request}
\title{Make an HTTP request}
\usage{
air_make_request(
  base,
  table_name,
  json_record_data,
  record_id = NULL,
  method = c("POST", "PATCH", "DELETE")
)
}
\arguments{
\item{base}{String. Base in airtable}

\item{table_name}{String. Table in airtable}

\item{json_record_data}{json or string. JSON formatted text with record data}

\item{record_id}{String or vector of strings. Record id}

\item{method}{String. One of "POST", "PATCH", or "DELETE"}
}
\value{
Status of HTTP request
}
\description{
Properly encodes HTTP requests
}
