% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_post.R
\name{air_post}
\alias{air_post}
\title{Get a list of records}
\usage{
air_post(
  base,
  table_name,
  limit = NULL,
  offset = NULL,
  view = NULL,
  fields = NULL,
  sortField = NULL,
  sortDirection = NULL,
  filterByFormula = NULL,
  combined_result = TRUE
)
}
\arguments{
\item{base}{String. Airtable base}

\item{table_name}{String. Table name}

\item{limit}{Numeric. (optional) A limit on the number of records to be returned.
Limit can range between 1 and 100.}

\item{offset}{Numeric. (optional) Page offset returned by the previous list-records
call. Note that this is represented by a record ID, not a numerical offset.}

\item{view}{String. (optional) The name or ID of the view}

\item{fields}{List. (optional) Only data for fields whose names are in this list
will be included in the records. Does not work when retrieving individual records with \code{record_id}}

\item{sortField}{String. (optional) The field name to use for sorting}

\item{sortDirection}{String. (optional) "asc" or "desc". The sort order in which the
records will be returned. Defaults to asc.}

\item{filterByFormula}{String. Use a formula to filter results. See \href{airtable docs}{https://support.airtable.com/hc/en-us/articles/223247187-How-to-sort-filter-or-retrieve-ordered-records-in-the-API}
this parameter to reduce the amount of data transferred.}

\item{combined_result}{Logical. If TRUE (default) all data is returned in the same data.
If FALSE table fields are returned in separate \code{fields} element.}
}
\value{
A data frame with records
}
\description{
Retrieve records where the request url would be over 16k characters (e.g. listing
many fields)
}
\details{
You can retrieve records in an order of a view by providing the name or ID of
the view in the view query parameter. The results will include only records
visible in the order they are displayed.
}
