% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_table_template}
\alias{air_table_template}
\title{Template for lists that describe tables in Airtable}
\usage{
air_table_template(table_name, description, fields_df)
}
\arguments{
\item{table_name}{String. Name of table}

\item{description}{String. Description of the table}

\item{fields_df}{Data frame. Data frame describing the field in a table.
Should contain a name, description,type, and options field.
if}
}
\value{
List with table name, description, and fields
}
\description{
Template for lists that describe tables in Airtable
}
