% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\name{air_update_data_frame}
\alias{air_update_data_frame}
\title{Update records from a dataframe}
\usage{
air_update_data_frame(base, table_name, record_ids, records)
}
\arguments{
\item{base}{String. Airtable base}

\item{table_name}{String. Table name}

\item{record_ids}{Vector of strings. Records to be modified}

\item{records}{Dataframe. Values to update}
}
\value{

}
\description{
Updates the values in a table by overwriting their current contents.
}
