% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_update_field}
\alias{air_update_field}
\title{Update a field name and/or description}
\usage{
air_update_field(base, table_id, field_id, name = NULL, description = NULL)
}
\arguments{
\item{base}{String. Base id}

\item{table_id}{String. ID for table that contains the field to be updated}

\item{field_id}{String. ID of field to be updated}

\item{name}{String. updated name (optional)}

\item{description}{String. updated description (option)}
}
\value{
List. Describes the changes that happened to the field
}
\description{
Must update either the name or the description.
See "https://airtable.com/developers/web/api/update-field" for more details.
}
\examples{
\dontrun{
base <- "appVjIfAo8AJlfTkx"

schema <- air_get_schema("appVjIfAo8AJlfTkx")

table_id <- schema$tables[1,c("id")]

field_id <- schema$tables$fields[[1]][2,]$id

## update name and description

name <- "New Name"

description <- "Updated Description"

out <- air_update_field(base = base,table_id = table_id,field_id = field_id,
name = name, description = description)

### just name

name <- "New New Name"

out <- air_update_field(base = base,table_id = table_id,field_id = field_id,
name = name)


## just description

description <- "Better description"

out <- air_update_field(base = base,table_id = table_id,field_id = field_id,
description = description)

## set name to number

name <- 1234

out <- air_update_field(base = base,table_id = table_id,field_id = field_id,
name = name)


# set description to number

description <- 1234

out <- air_update_field(base = base,table_id = table_id,field_id = field_id,
description = description)

# bulk update names and descriptions from a data frame

field_ids <- schema$tables$fields[[1]]$id

field_names <- sprintf("\%s_bulk_update",schema$tables$fields[[1]]$name)

field_descriptions <- sprintf("\%s BULK UPDATE",
schema$tables$fields[[1]]$description)

df <- data.frame("field_id"= field_ids,"name"=field_names,
"description"=field_descriptions)

purrr::pmap(df,function(field_id,name,description){
  air_update_field(base = base,table_id = table_id,field_id = field_id,
  name = name, description = description)
})
}
}
