% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_update_metadata_table}
\alias{air_update_metadata_table}
\title{Update the structural metadata table}
\usage{
air_update_metadata_table(
  base,
  meta_data,
  table_name = "Meta Data",
  join_field = "field_id",
  record_id_field = "id"
)
}
\arguments{
\item{base}{String. Base id}

\item{meta_data}{Data frame. Contains metadata records. From air_generate_metadata*}

\item{table_name}{String. Name of metadata table}

\item{join_field}{String. Name of field to join new and current metadata. Likely \code{field_id}}

\item{record_id_field}{String. Name of record id field. Like \code{id}}
}
\description{
Update the structural metadata table
}
\examples{
\dontrun{
base = "appVjIfAo8AJlfTkx"
metadata <- air_generate_metadata_from_api(base = base)
air_update_metadata_table(base,metadata)
}

}
