% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/airtable.R
\name{Airtable}
\alias{Airtable}
\title{Class: Airtable}
\usage{
Airtable(app, tables)
}
\arguments{
\item{app}{A length-one character vector for app name}

\item{tables}{A character vector for tables}
}
\value{
An Airtable wrapper object
}
\description{
\code{Airtable} is an S3 class to represent an Airtbale API wrapper
}

