% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{airtable_key}
\alias{airtable_key}
\title{Get airtable key if exists
otherwise set airtable key}
\usage{
airtable_key(force = FALSE)
}
\arguments{
\item{force}{A logical}
}
\value{
A character
}
\description{
Get airtable key if exists
otherwise set airtable key
}

