% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{airtable_PATCH}
\alias{airtable_PATCH}
\title{Update some(but not all) fields of a table record}
\usage{
airtable_PATCH(record_endpoint, fields)
}
\arguments{
\item{record_endpoint}{A length-one character vector}

\item{fields}{A list}
}
\description{
Update a table record by issuing a PATCH request to the record endpoint.
Note that any fields that are not included will not be updated.
}

