% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{airtable_POST}
\alias{airtable_POST}
\title{Create a new record}
\usage{
airtable_POST(table_endpoint, fields)
}
\arguments{
\item{table_endpoint}{A length-one character vector. Path to a table in your Airtable account. e.g "appANrRXq7xaOU0dd/table_name"}

\item{fields}{A list}
}
\description{
Create a new record by issuing POST request to a table endpoint
}

