% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/compose_endpoint.R
\name{compose_endpoint}
\alias{compose_endpoint}
\title{Compose endpoint to a table or a table record}
\usage{
compose_endpoint(base, table, record_id = NULL)
}
\arguments{
\item{base}{An Airtable wrapper object}

\item{table}{A length-one character vector}

\item{record_id}{An optional length-one character vector for record id e.g "recbYejyoIyRBloTj"}
}
\value{
A length-one character vector which represents the composed endpoint
}
\description{
\code{compose_endpoint} composes an endpoint to table or a table record
}

