% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/compose_request_arguments.R
\name{compose_request_arguments}
\alias{compose_request_arguments}
\title{Compose request arguments}
\usage{
compose_request_arguments(base, table, resource_id = NULL, fields = NULL,
  method, query = NULL)
}
\arguments{
\item{base}{A list}

\item{table}{A length-one character vector}

\item{resource_id}{An optional length-one character vector}

\item{fields}{An optional list}

\item{method}{A length-one character vector. e.g "put", "patch", ...}

\item{query}{An optional list}
}
\value{
A list
}
\description{
\code{compose_request_arguments} composes \code{httr} request arguments as a list
}

