% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/create_record.R
\name{create_record}
\alias{create_record}
\title{Create a record to Airtalbe}
\usage{
create_record(base, table, fields)
}
\arguments{
\item{base}{A list}

\item{table}{A length-one character vector}

\item{fields}{A list}
}
\value{
A request object
}
\description{
\code{create_record} creates an Airtable record remotely by issuing POST request to the table endpoint
}

