% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/delete_record.R
\name{delete_record}
\alias{delete_record}
\title{Delete a record}
\usage{
delete_record(air_options, table, record_id)
}
\arguments{
\item{air_options}{A list}

\item{table}{A length-one character vector}

\item{record_id}{A length-one character vector}
}
\description{
\code{delete_record} deletes a record by issuing DELETE request to a record endpoint.
}

