% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/dispatch_request.R
\name{dispatch_request}
\alias{dispatch_request}
\title{Dispatch a HTTP request for Airtable API}
\usage{
dispatch_request(airtable, resource_id = NULL, fields = NULL, method,
  query = NULL)
}
\arguments{
\item{airtable}{An Airtable wrapper object}

\item{resource_id}{An optional length-one character vector. The default is NULL.}

\item{fields}{An optional list to dispatch(as JSON object)}

\item{method}{A length-one character vector. e.g "put", "patch", "post"}

\item{query}{An optional list}
}
\value{
A request object
}
\description{
Dispatch a HTTP request for Airtable API
}

