% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_all.R
\name{fetch_all}
\alias{fetch_all}
\title{Fetch All Records in an Airtable}
\usage{
fetch_all(base, table_name, ...)
}
\arguments{
\item{base}{String. ID for the base or app to be fetched}

\item{table_name}{String. Name of the table to be fetched from the base}

\item{...}{Additional arguments to pass to \code{air_get}. \code{view} is a
commonly used additional argument.}
}
\value{
dataframe
}
\description{
Airtable limits the number of records that can be pulled from a base to 100.
This function pulls records based on a query then checks if there is an
offset value. While there is an offset value, it uses that value to generate
the next query, thus moving down the records until all records have been
fetched from the database.
}
\section{Adding airtable API key to your environment}{

Airtable requires an api key to fetch data.
Generate the airtable API key from your [Airtable account](http://airtable.com/account) page.

\code{airtabler} functions will read the API key from
environment variable \code{AIRTABLE_API_KEY}. To start R session with the
initialized environment variable create an \code{.Renviron} file in your
home directory with a line like this:

\code{AIRTABLE_API_KEY=your_api_key_here}

You can use \code{usethis::edit_r_environ()} to edit your find and edit your
file.
}

\examples{
# Each base a fully described API
app_id <- "appwlxIzmQx5njRtQ" # ID for the base we are fetching.
# Note that you can pass a `view` argument to air_get or fetch_all to get only
# a view of a table (say, only validated records, or some other filtered view),
# e.g.,
# bats <- fetch_all(app_id, "Bat", view = "Validated Records")
talks <- fetch_all(app_id, "TALKS")

}
