% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_get_json.R
\name{fetch_all_json}
\alias{fetch_all_json}
\title{Get the full outputs of a table as single json object}
\usage{
fetch_all_json(base, table_name, ...)
}
\arguments{
\item{base}{String. Base ID}

\item{table_name}{String. Table name}

\item{...}{additional parameters to pass to air_get_json}
}
\value{
json as string
}
\description{
Get the full outputs of a table as single json object
}
