% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\name{get_offset}
\alias{get_offset}
\title{Get offset}
\usage{
get_offset(x)
}
\arguments{
\item{x}{Last result}
}
\value{
Airtable offset id
}
\description{
Returns airtable offset id from previous select
}

