% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_field_values.R
\name{get_unique_field_values}
\alias{get_unique_field_values}
\title{Get unique values from a field}
\usage{
get_unique_field_values(base, table_name, fields)
}
\arguments{
\item{base}{String. ID of airtable base}

\item{table_name}{String. Name of table in base}

\item{fields}{List. Names of fields}
}
\value{
vector of unique values
}
\description{
Because the airtable api lacks an easy solution for getting
unique values from a field in a table, we have this function
which accepts a list of fields and returns their unique
values. Currently, if multiple fields are listed, all unique
values from all fields will be returned in a single vector.
This may change in future iterations.
}
