% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_url.R
\name{read_excel_url}
\alias{read_excel_url}
\title{Read an excel file from URL}
\usage{
read_excel_url(url, fileext = ".xslx", ...)
}
\arguments{
\item{url}{String. Url for file}

\item{fileext}{String. File extension for temp file}

\item{...}{additional arguments to pass to \code{read_excel}}
}
\value{
tibble
}
\description{
Extends \code{readxl::read_excel} to allow for reading from a URL.
}
