% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{request_arguments}
\alias{request_arguments}
\title{Get common arguments for POST, PATCH, PUT request}
\usage{
request_arguments(path, body)
}
\arguments{
\item{path}{}

\item{body}{}
}
\description{
Get common arguments for POST, PATCH, PUT request
}

