% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/retrieve_record.R
\name{retrieve_record}
\alias{retrieve_record}
\title{Retrieve a record from Airtable}
\usage{
retrieve_record(airtable, record_id)
}
\arguments{
\item{airtable}{An Airtable wrapper object}

\item{record_id}{A length-one character vector for record id. e.g "recbYejyoIyRBloTj"}
}
\value{
A list which represents the retrieved record
}
\description{
\code{retrieve_record} returns a record by issuing GET request to a record endpoint
}

