% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{set_diff}
\alias{set_diff}
\title{Get items that differ between x and y}
\usage{
set_diff(x, y)
}
\arguments{
\item{x}{a set of values.}

\item{y}{a set of values.}
}
\value{
Unique values from X and Y, NULL if no unique values.
}
\description{
Unlike setdiff, this function creates the union of x and y then
removes values that are in the intersect, providing values
that are unique to X and values that are unique to Y.
}
\examples{
a <- 1:3
b <- 2:4

set_diff(a,b)
# returns 1,4

x <- 1:3
y <- 1:3

set_diff(x,y)
# returns NULL

}
