% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/update_record.R
\name{update_record}
\alias{update_record}
\title{Update a record}
\usage{
update_record(airtable, record_id, fields, method = "PATCH")
}
\arguments{
\item{airtable}{An Airtable wrapper object}

\item{record_id}{A length-one character vector}

\item{fields}{A list for fields}

\item{method}{A length-one character vector. The default is "PATCH"}
}
\value{
A request object
}
\description{
\code{update_record} updates a record by issuing PATCH or PUT request to a record endpoint.
Note that if you use PUT method, any fields that are not included will be cleared.
}

