% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_create_table}
\alias{air_create_table}
\title{A function to create new tables in a base}
\usage{
air_create_table(base, table_list)
}
\arguments{
\item{base}{String. ID for the base}

\item{table_list}{List. see \code{air_table_template}}
}
\value{
Data frame of table schema
}
\description{
Takes a list object with appropriate arguments (see \code{air_table_template})
converts it to JSON then adds it to the specified base.
}
\note{
See https://airtable.com/developers/web/api/create-table
}
\examples{

\dontrun{
base <- "appQ94sELAtFnXPxx"

base_schema <- air_get_schema(base)

tables<- base_schema$tables

field_names <- c("Planet","Chapter","Book", "Known Inhabitants")

field_desc <- c("Name of planet in Foundation Series",
                "Chapters where planet is referenced",
                "Books where planet is referenced",
                "Characters mentioned as living on or being from that planet")

field_types <- c("singleLineText",rep("multipleRecordLinks",3))

field_options <- c(NA,list(
  list(
    linkedTableId = tables[tables$name == "Chapter","id"]
  )
),
list(
  list(
    linkedTableId = tables[tables$name == "Book","id"]
  )
),
list(
  list(
    linkedTableId = tables[tables$name == "Character","id"]
  )
)
)

field_df<- air_fields_df_template(name = field_names,
                                  description = field_desc,
                                  type = field_types,
                                  options = field_options)

table_list <- air_table_template(table_name = "Planet",
                                 description = "Planets of Foundation",
                                 fields_df = field_tables)

air_create_table(base, table_list)

}
}
