% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_dump}
\alias{air_dump}
\title{Output tables to a specified file format}
\usage{
air_dump(
  base = Sys.getenv("airtable_metadata_base"),
  table_name_metadata = "meta data",
  table_name_description = "description",
  all_table_names = NULL,
  output_dir = "outputs",
  output_id = rlang::hash(Sys.time()),
  output_format = "csv",
  to_frictionless_data = FALSE
)
}
\arguments{
\item{base}{String. ID for your base from Airtable. Generally 'appXXXXXXXXXXXXXX'}

\item{table_name_metadata}{String. Name of structural metadata table. PROVIDE LINK TO DEF OF STRUCTURAL METADATA}

\item{table_name_description}{String. Name of the descriptive metadata table. PROVIDE LINK TO DEF OF DESCRIPTIVE METADATA}

\item{all_table_names}{Vector. Do not use if providing table_name_metadata.
A character vector of all table names in your base. I.e c("table 1","table 2").}

\item{output_dir}{String. Folder containing outputs. If not specified, a folder
with the name of the output_id will be used.}

\item{output_id}{String. ID used to uniquely identify outputs. A folder with
this name will be created in output_dir.}

\item{output_format}{String. Default is csv. Options include: csv}

\item{to_frictionless_data}{Logical. Requires table_name_metadata and table_name_description.
Should data be output as a frictionless data package. (To be Added)}
}
\value{
List of file paths created.
}
\description{
This function uses metadata tables in your base to extract all
base table to a different file format. Currently only CSV is supported.
For information about creating metadata tables in your base see the
\href{https://ecohealthalliance.github.io/eha-ma-handbook/8-airtable.html#managing-data}{EHA MA Handbook}
}
