% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_dump_to_csv}
\alias{air_dump_to_csv}
\title{Save air_dump output to csv}
\usage{
air_dump_to_csv(
  table_list,
  output_dir = "outputs",
  attachments_dir = NULL,
  overwrite = FALSE,
  output_id = NULL
)
}
\arguments{
\item{table_list}{List. List of data.frames output from \code{air_dump}}

\item{output_dir}{String. Folder containing output files}

\item{attachments_dir}{String. What folder are base attachments stored in?}

\item{overwrite}{Logical. Should outputs be overwritten if they already exist?}

\item{output_id}{String. Optional identifier for the data set - if NULL an
ID will be generated using a hash of the data.}
}
\value{
Vector of file paths
}
\description{
Saves data.frames from air_dump to csv files. File names are determined by
the names of the list objects from air_dump. Files will be saved in folder
with a unique name, inside the folder specified by \code{output_dir}. The
unique name is generated from a hash of the air_dump output.
}
