% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_fields_df_template}
\alias{air_fields_df_template}
\title{Template for for creating a table from a tibble}
\usage{
air_fields_df_template(name, description, type, options = NA)
}
\arguments{
\item{name}{String. Names of fields in the table}

\item{description}{String. Descriptions of fields}

\item{type}{String. Type of columns. For values see \url{https://airtable.com/developers/web/api/model/field-type}}

\item{options}{List. Options will be converted from lists to JSON. For field options see \url{https://airtable.com/developers/web/api/field-model}}
}
\value{
Tibble with attributes required for fields in a table
}
\description{
Convenience function for creating the content of tables that will created or
updated viaAPI.
}
\examples{

\dontrun{
base <- "appQ94sELAtFnXPxx"

base_schema <- air_get_schema(base)

tables<- base_schema$tables

field_names <- c("Planet","Chapter","Book", "Known Inhabitants")

field_desc <- c("Name of planet in Foundation Series",
                "Chapters where planet is referenced",
                "Books where planet is referenced",
                "Characters mentioned as living on or being from that planet")

field_types <- c("singleLineText",rep("multipleRecordLinks",3))

field_options <- c(NA,list(
  list(
    linkedTableId = tables[tables$name == "Chapter","id"]
  )
),
list(
  list(
    linkedTableId = tables[tables$name == "Book","id"]
  )
),
list(
  list(
    linkedTableId = tables[tables$name == "Character","id"]
  )
)
)

field_df<- air_fields_df_template(name = field_names,
                                  description = field_desc,
                                  type = field_types,
                                  options = field_options)

table_list <- air_table_template(table_name = "Planet",description = "Planets of Foundation",fields_df = field_tables)

air_create_table(base, table_list)
}
}
