% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_generate_metadata_from_api}
\alias{air_generate_metadata_from_api}
\title{Generate structural metadata from the api}
\usage{
air_generate_metadata_from_api(
  base,
  metadata_table_name = "Meta Data",
  include_metadata_table = FALSE,
  field_names_to_snake_case = TRUE
)
}
\arguments{
\item{base}{String. Base id}

\item{metadata_table_name}{String. Name of exisiting structural metadata table if it exists}

\item{include_metadata_table}{Logical. Should the structural metadata table be included in the metadata?}

\item{field_names_to_snake_case}{Logical. Should the field names in the metadata table be snake_case?}
}
\value{
A data frame with metadata
}
\description{
Structural metadata describes the contents of your base and how they are linked.
The structural metadata are created from the base schema. The nested schema
structure is flattened into a more user-friendly table which can then be
inserted as a table into the base with \code{air_created_metadata_table} and/or
used in a data export with \code{air_dump}.
}
\details{
This function requires that the api token has the ability to read
the base schema.
}
\examples{

\dontrun{

base <- "appXXXXXXXX"
metadata  <- air_generate_metadata_from_api(base)

}
}
