% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_get_metadata_from_table}
\alias{air_get_metadata_from_table}
\title{Pull the metadata table from Airtable}
\usage{
air_get_metadata_from_table(
  base,
  table_name,
  add_id_field = TRUE,
  field_names_to_snakecase = TRUE
)
}
\arguments{
\item{base}{String. ID for your base from Airtable. Generally 'appXXXXXXXXXXXXXX'}

\item{table_name}{String. Name of structural metadata table - the metadata that
describes how tables and fields fit together.}

\item{add_id_field}{Logical. If true, an "id" field is added to each table}

\item{field_names_to_snakecase}{Logical. If true, values in the field_names
column are converted to snake_case}
}
\value{
data.frame with metadata table
}
\description{
Airtable allows all users to access the metadata API.
The recommended workflow for creating this table is to use
air_generate_metadata_from_api to extract the structural metadata from the base
schema and then use air_create_metadata_table to add the table to your
base.
}
\details{
For information about creating metadata tables in your base see the
\href{https://ecohealthalliance.github.io/eha-ma-handbook/8-airtable.html#managing-data}{EHA MA Handbook}

Requires the following fields: table_name, field_name
}
