% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_get_schema.R, R/air_metadata_api.R
\name{air_get_schema}
\alias{air_get_schema}
\title{Get base schema}
\usage{
air_get_schema(base, ...)

air_get_schema(base, ...)
}
\arguments{
\item{base}{Airtable base ID}

\item{...}{additional paramters}
}
\value{
list of schema

list of schema
}
\description{
Get the schema for the tables in a base. This is a wrapper for the api call
Get base schema.

Get the schema for the tables in a base. This is a wrapper for the api call
Get base schema.
}
\section{Using Metadata API}{

Metadata api is currently available to all users.


Metadata api is currently available to all users.
}

