% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_get_schema.R
\name{air_get_schema}
\alias{air_get_schema}
\title{Get base schema - enterprise only}
\usage{
air_get_schema(base, ...)
}
\arguments{
\item{base}{Airtable base ID}

\item{...}{additional paramters}
}
\value{
list of schema
}
\description{
Metadata API currently only available via enterprise accounts.
Get the schema for the tables in a base.
}
\section{Using Metadata API}{

The meta data api is not accepting new requests for client secrets as of
06 July 2021. Will update when metadata api becomes available.
}

