% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_metadata_api.R
\name{air_list_bases}
\alias{air_list_bases}
\title{Get list of bases for an Token}
\usage{
air_list_bases(request_url = "https://api.airtable.com/v0/meta/bases")
}
\arguments{
\item{request_url}{String. URL for api endpoint}
}
\value{
list. List of bases a token can access.
}
\description{
Each token you provisision is given access to a certain set of bases or
workspaces. This function lists all bases associated with a token.
}
\examples{

\dontrun{
air_list_bases()
}

}
