% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_dump.R
\name{air_update_description_table}
\alias{air_update_description_table}
\title{Update the description table}
\usage{
air_update_description_table(
  base,
  description,
  table_name = "Description",
  join_field = "title",
  record_id_field = "id"
)
}
\arguments{
\item{base}{String. Base id}

\item{description}{Data frame. Contains updated description}

\item{table_name}{String. Name of description table}

\item{join_field}{String. Field to perform join on}

\item{record_id_field}{String. Name of the record id field}
}
\value{
list that logs updates
}
\description{
Update the descriptive metadata table in airtable
}
