% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\name{airtable}
\alias{airtable}
\title{Get airtable base object}
\usage{
airtable(base, tables)
}
\arguments{
\item{base}{Airtable base}

\item{tables}{Table names in the airtable base (character vector)}
}
\value{
Airtable base object with elements named by table names.
  Each element contains functions
  \item{get}{returns table records, see \code{\link{air_get}} for details}
  \item{insert}{insert table record, see \code{\link{air_insert}} for details}
  \item{update}{updates table record, see \code{\link{air_update}} for details}
  \item{delete}{deletes table record, see \code{\link{air_delete}} for details}
}
\description{
Creates airtable object with tables and functions
}
\examples{
\dontrun{
TravelBucketList <-
  airtable(
    base = "the_base_id",
    tables = c("Destinations", "Hotels", "Travel Partners")
  )
  hotels <- TravelBucketList$Hotels$get()
  destinations <- TravelBucketList$Destinations$get()
}
}
