% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{airtable_GET}
\alias{airtable_GET}
\title{List records}
\usage{
airtable_GET(table_endpoint, record = NULL, query = NULL)
}
\arguments{
\item{table_endpoint}{A length-one character vector. Path to a table in your Airtable account. e.g "appANrRXq7xaOU0dd/table_name"}

\item{query}{A list. Query string parameters. e.g list(page = 2)}
}
\description{
List records or retreive a particular record by issuing GET request to a table/record endpoint
}

