% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{airtable_GET}
\alias{airtable_GET}
\title{Get data from Airtable}
\usage{
airtable_GET(path, query = NULL)
}
\arguments{
\item{path}{A length-one character vector. Path to a table in your Airtable account. e.g "appANrRXq7xaOU0dd/table_name"}

\item{query}{A list. Query string parameters. e.g list(page = 2)}
}
\description{
Get records from Airtable API then convert them into a list.
}

