% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/functions.R
\name{airtable_key}
\alias{airtable_key}
\title{Get Airtable API key if exists
otherwise set Airtable API key}
\usage{
airtable_key(force = FALSE)
}
\arguments{
\item{force}{An optional length-one logical vector}
}
\value{
A length-one character vecotr
}
\description{
Get Airtable API key if exists
otherwise set Airtable API key
}

