% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\docType{package}
\name{airtabler-package}
\alias{airtabler}
\alias{airtabler-package}
\title{airtabler: Interface to the Airtable API}
\description{
Provides access to the Airtable API (\url{http://airtable.com/api}).
}
\details{
After you've created and configured the schema of an Airtable base from the
graphical interface, your Airtable base will provide its own API to create,
read, update, and destroy records (from \url{http://airtable.com/api})
}
\section{API key}{
 \pkg{airtabler} functions will read the API key from
  environment variable \code{AIRTABLE_API_KEY}. To start R session with the
  initialized environvent variable create an .Renviron file in your R home
  with a line like this:

  \code{AIRTABLE_API_KEY=************}

  To check where your R home is, try \code{normalizePath("~")}.
}

\section{Usage}{

  Use \code{\link{air_get}}, \code{\link{air_insert}},
  \code{\link{air_update}} and \code{\link{air_delete}} to access your
  airtable data.
}

