% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtabler.R
\docType{package}
\name{airtabler-package}
\alias{airtabler-package}
\alias{airtabler}
\title{airtabler: Interface to the Airtable API}
\description{
Provides access to the Airtable API (\url{http://airtable.com/api}).
}
\section{Setup}{

  Create and configure the schema of an Airtable base on (\url{http://airtable.com})
  and check the API on \url{http://airtable.com/api}.
}

\section{API key}{

  Generate the airtable API key from your Airtable account page
  (http://airtable.com/account).

  \pkg{airtabler} functions will read the API key from
  environment variable \code{AIRTABLE_API_KEY}. To start R session with the
  initialized environment variable create an \code{.Renviron} file in your R home
  with a line like this:

  \code{AIRTABLE_API_KEY=************}

  To check where your R home is, try \code{normalizePath("~")}.
}

\section{Usage}{

  Use \code{\link{airtable}} function to get airtable base object
  or just call primitives \code{\link{air_get}}, \code{\link{air_insert}},
  \code{\link{air_update}} and \code{\link{air_delete}} to access your
  airtable data.
}

