% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/list_records.R
\name{list_records}
\alias{list_records}
\title{List records form Airtable}
\usage{
list_records(base, table, offset = NULL, filterByFormula = NULL)
}
\arguments{
\item{base}{A list}

\item{table}{A length-one character vector}

\item{offset}{An optional length-one character vector}

\item{filterByFormula}{An optional length-one character vector}
}
\description{
\code{list_records} returns records by issuing GET request to a table endpoint
}

